Fix previous commit, move comment.proto and suite.js to tests directory.
